Separate importers from backend.
Separating importers from backend allows to have them running
remotely independently from the backend serving the UI.

An programmatic API is also provided to easily create a new
importer in Go.

Currently, importers are directly connected to the database but
in the long-term the backend should act as a gateway to control
DB accesses.