Fix issue with DB ID not found in generator.
In the case there is a relation in the projection, but the related
assets are not in the projection, the assets are not pushed into the
registry leading to missing IDs when searching into it. In that case we
can simply return id -1 because the relation won't be displayed anyway
as we are missing one or both vertices.