Ensure good use of db transactions and remove duplicate constraint in schema.
Removing the duplicate constraint avoid some issues with the transaction
logcks and is supposed to speed up inserts. The duplication constraints
is still partially checked since the id which is the
hash of the type and value and it's the primary key of the table.

It does not prevent a bug in the code to generate duplicate with different
IDs but this can be checked regularly and offline.