Add streaming support on read side.
Streaming was implemented only on the write side but when multiple data
sources requested a graph from the API, huge graph could have been
loaded in memory before being encoded and sent over the wire. With this
encoder, each relation or asset is itself encoded and sent independently
and the whole graph is reconstructed on client side.

The encoding format is definitely not optimized yet since it uses json
format and duplicated assets could be sent multiple times but that is
worth it considering the amount of memory saved. The deduplication is
done on the client side anyway.