Add a way to return sources from the API.
This is a first approach to return sources of items but not the right
one since counting items does not work anymore. A better approach is to
split the queries in two parts: the actual graph query and then another
separate one to retrieve the sources for earch item. That way we get the
proper count in the first query and we get the sources from the second.
Everything could be handled by the backend or exposed as an API.