fix(tx): do not consider assets from relations as present in the graph
When creating a transaction we load the existing assets and relations
to perform a diff and push the updates. We mistakenly considered that
assets coming from relation were part of the graph for the source.
However this is not necessarily true when there are errors in the graph.
We now only consider assets explicitely listed in order to repair the
graph if some data is missing.